Revert "Remove superfluous test class hierarchy"
This reverts commit b64befc92a8b59b211d16e5d5d82ccdfb2e5462f.

and move the tests to the same place
