+++ kafka/src/test/groovy/io/micronaut/configuration/kafka/errors/KafkaPartitionedErrorsSpec.groovy
@@ -6,10 +6,6 @@ import spock.lang.Retry
 
 import java.util.stream.IntStream
 
-import static io.micronaut.configuration.kafka.annotation.ErrorStrategyValue.RESUME_AT_NEXT_RECORD
-import static io.micronaut.configuration.kafka.annotation.ErrorStrategyValue.RETRY_ON_ERROR
-import static io.micronaut.configuration.kafka.annotation.OffsetReset.EARLIEST
-
 @Stepwise
 @Retry
 @Ignore("https://github.com/micronaut-projects/micronaut-kafka/issues/514")
@@ -22,10 +18,7 @@ class KafkaPartitionedErrorsSpec extends KafkaErrorsSpec {
 
     @Override
     protected Map<String, Object> getConfiguration() {
-        super.configuration + [
-                'kafka.consumers.default.max.poll.records': 10,
-                'spec.name'                               : KafkaErrorsSpec.class.simpleName
-        ]
+        return super.getConfiguration() + ['spec.name': KafkaErrorsSpec.class.simpleName]
     }
 
     void "should correctly handle error strategies"() {
@@ -58,124 +51,13 @@ class KafkaPartitionedErrorsSpec extends KafkaErrorsSpec {
 
     void "should be correctly processing in partitions"() {
         expect:
-        conditions.eventually {
-            listenerWithErrorStrategyResumeAtNextRecord.partitions.size() == 3
-            listenerWithErrorStrategyRetryOnError.partitions.size() == 3
-            listenerWithErrorStrategyRetryOnError10Times.partitions.size() == 3
-            listenerSyncPerRecordWithErrorStrategyRetryOnError10Times.partitions.size() == 3
-            listenerWithErrorStrategyNone.partitions.size() == 3
-        }
-    }
-
-
-    @Shared
-    TestListenerWithErrorStrategyResumeAtNextRecord listenerWithErrorStrategyResumeAtNextRecord
-
-    @Shared
-    TestListenerWithErrorStrategyRetryOnError listenerWithErrorStrategyRetryOnError
-
-    @Shared
-    TestListenerWithErrorStrategyRetryOnError10Times listenerWithErrorStrategyRetryOnError10Times
-
-    @Shared
-    TestListenerWithErrorStrategyNone listenerWithErrorStrategyNone
-
-    @Shared
-    TestListenerSyncPerRecordWithErrorStrategyRetryOnError10Times listenerSyncPerRecordWithErrorStrategyRetryOnError10Times
-
-    @Shared
-    TestProducer producer
-
-    @Override
-    protected int getConditionsTimeout() {
-        return 120
-    }
-
-    void setupSpec() {
-        listenerWithErrorStrategyResumeAtNextRecord = context.getBean(TestListenerWithErrorStrategyResumeAtNextRecord)
-        listenerWithErrorStrategyRetryOnError = context.getBean(TestListenerWithErrorStrategyRetryOnError)
-        listenerWithErrorStrategyRetryOnError10Times = context.getBean(TestListenerWithErrorStrategyRetryOnError10Times)
-        listenerWithErrorStrategyNone = context.getBean(TestListenerWithErrorStrategyNone)
-        listenerSyncPerRecordWithErrorStrategyRetryOnError10Times = context.getBean(TestListenerSyncPerRecordWithErrorStrategyRetryOnError10Times)
-        producer = context.getBean(TestProducer)
-    }
-
-    @Introspected
-    @EqualsAndHashCode
-    static class TestEvent implements Comparable<TestEvent> {
-
-        final int count
-
-        TestEvent(int count) {
-            this.count = count
-        }
-
-        @Override
-        String toString() {
-            count
-        }
-
-        @Override
-        int compareTo(@NotNull TestEvent o) {
-            return count <=> o.count
-        }
-    }
-
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaListener(offsetReset = EARLIEST, errorStrategy = @ErrorStrategy(value = RESUME_AT_NEXT_RECORD))
-    static class TestListenerWithErrorStrategyResumeAtNextRecord extends AbstractTestListener {
-    }
-
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaListener(offsetReset = EARLIEST, errorStrategy = @ErrorStrategy(value = RETRY_ON_ERROR))
-    static class TestListenerWithErrorStrategyRetryOnError extends AbstractTestListener {
-    }
-
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaListener(offsetReset = EARLIEST, errorStrategy = @ErrorStrategy(value = RETRY_ON_ERROR, retryCount = 10))
-    static class TestListenerWithErrorStrategyRetryOnError10Times extends AbstractTestListener {
-    }
-
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaListener(offsetReset = EARLIEST)
-    static class TestListenerWithErrorStrategyNone extends AbstractTestListener {
-    }
-
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaListener(offsetReset = EARLIEST, offsetStrategy = OffsetStrategy.SYNC_PER_RECORD, errorStrategy = @ErrorStrategy(value = RETRY_ON_ERROR, retryCount = 10))
-    static class TestListenerSyncPerRecordWithErrorStrategyRetryOnError10Times extends AbstractTestListener {
-    }
-
-    @Slf4j
-    static abstract class AbstractTestListener implements KafkaListenerExceptionHandler {
-
-        TreeSet<Integer> partitions = []
-        List<TestEvent> failed = []
-        TreeSet<TestEvent> events = []
-        List<KafkaListenerException> exceptions = []
-
-        @Topic("test-topic")
-        void receive(@KafkaKey UUID key, @KafkaPartition int partition, TestEvent event) {
-            partitions << partition
-            if (event.count == 3) {
-                failed << event
-                throw new IllegalArgumentException("BOOM")
+            conditions.eventually {
+                listenerWithErrorStrategyResumeAtNextRecord.partitions.size() == 3
+                listenerWithErrorStrategyRetryOnError.partitions.size() == 3
+                listenerWithErrorStrategyRetryOnError10Times.partitions.size() == 3
+                listenerSyncPerRecordWithErrorStrategyRetryOnError10Times.partitions.size() == 3
+                listenerWithErrorStrategyNone.partitions.size() == 3
             }
-//            System.out.println(partition + " " + event + " " + this)
-            events << event
-        }
-
-        @Override
-        void handle(KafkaListenerException exception) {
-            exceptions << exception
-        }
     }
 
-    @Requires(property = 'spec.name', value = 'KafkaErrorsSpec')
-    @KafkaClient
-    static interface TestProducer {
-
-        @Topic("test-topic")
-        void send(@KafkaKey UUID key, TestEvent event);
-    }
 }
